Move picks conversion functions into tree view
This will allow them to be used elsewhere
